requirements.txt
streamlit>=1.23.0
huggingface-hub>=0.14.1
python-dotenv